requirements.txt
faiss-cpu
sentence-transformers
streamlit
pandas
requests
nltk
